requirements.txt
matplotlib
numpy
opencv-python
Pillow
PyYAML
scikit-image
scipy
seaborn
streamlit
streamlit-webrtc
thop
torch
torchvision
tqdm